% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ensure_that}
\alias{ensure_that}
\title{Ensure certain conditions for a value at runtime.}
\usage{
ensure_that(value., ...)
}
\arguments{
\item{value.}{The value which is to be ensured.}

\item{...}{Conditions which must pass for the ensurance to be fulfilled.
Any named argument will treated as values available when evaluating the
conditions. To reference the value itself use the dot-placeholder, `.`}
}
\value{
value.
}
\description{
Ensure certain conditions for a value at runtime.
}
\details{
if value. does not comply with conditions an error is raised.
}
\examples{
\dontrun{
library(magrittr)

# Will pass
matrix(runif(16), 4, 4) \%>\%
	 ensure_that(ncol(.) == nrow(.), all(. <= 1))

# Will raise an error
matrix(NA, 4, 4) \%>\%
  ensure_that(. \%>\% anyNA \%>\% not)
}
}

