% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ensuring}
\alias{ensuring}
\title{Make a contract ensuring certain properties of a value.}
\usage{
ensuring(...)
}
\arguments{
\item{...}{conditions which must pass for the ensuring contract to be
fulfilled. Any named argument will treated as values available
when evaluating the conditions. To reference the value itself
use the dot-placeholder, `.`}
}
\value{
an ensuring function
}
\description{
Make a contract ensuring certain properties of a value.
}
\examples{
\dontrun{
library(magrittr)

# Ensures a matrix is square. Read `<-` as "is".
ensure_square <- ensuring(NCOL(.) == NROW(.))

diag(4) \%>\%
  ensure_square
}
}

