% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/function_.R
\name{function_}
\alias{function_}
\title{Define a type-safe function.}
\usage{
function_(...)
}
\arguments{
\item{...}{Arguments for the function and the function body as the last
  argument.}
}
\value{
a type-safe function
}
\description{
Using \code{function_} is an alternative way to define a function where
arguments are type-safe. Arguments are specified as \code{name ~ type} or as
\code{type:name ~ value}. The last argument to \code{function_} is the body
of the function. If no type is specified, it will default to
\code{some}, which is true for anything.
}
\examples{
f <- function_(a ~ integer, b ~ character: "Hello, World", {
  rep(b, a)
})
}

