% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse_argument.R
\name{parse_argument}
\alias{parse_argument}
\title{Parse a \code{function_} argument definition.}
\usage{
parse_argument(arg)
}
\arguments{
\item{arg}{A language object representing the argument.}
}
\value{
list with the following entries:
\itemize{
  \item \code{type} character: the type
  \item \code{name} character: the name
  \item \code{has_default} logical: whether a default value was specified
  \item \code{default} the default value (or NULL if not specified.)
}
}
\description{
An argument definition as passed to \code{function_} is checked and
parsed. The argument is expected to have the syntax \code{name ~ type: default}
where the \code{: default} is optional specification of the default value.
}
\details{
Note that since a default value can be \code{NULL}, it is necessary to
check whether the default was specified in the argument, see the return value.
}

