% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensures_that.R
\name{print.ensurer}
\alias{print.ensurer}
\title{Print method for ensurer contracts}
\usage{
\method{print}{ensurer}(x, ...)
}
\arguments{
\item{x}{a function made with \code{ensures_that}}

\item{...}{not used.}
}
\value{
x
}
\description{
Print method for ensurer contracts
}

