% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/types.R
\name{type_character}
\alias{type_character}
\title{Whole number}
\usage{
type_character(.)
}
\arguments{
\item{.}{Object to be tested for being of character type.}
}
\value{
identity
}
\description{
Whole number
}

