% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/types.R
\name{type_integer}
\alias{type_integer}
\title{Integer type}
\usage{
type_integer(.)
}
\arguments{
\item{.}{Object to be tested for being of integer type.}
}
\value{
identity
}
\description{
Integer type
}

