% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/types.R
\name{type_numeric}
\alias{type_numeric}
\title{Numeric type}
\usage{
type_numeric(.)
}
\arguments{
\item{.}{Object to be tested for being of numeric type.}
}
\value{
identity
}
\description{
Numeric type
}

