% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/types.R
\name{type_whole}
\alias{type_whole}
\title{Whole number type}
\usage{
type_whole(.)
}
\arguments{
\item{.}{Object to be tested for being a whole number.}
}
\value{
identity
}
\description{
Whole number type
}

