% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ensures_that}
\alias{ensures_that}
\title{Make a contract which ensures that certain conditions are fulfilled.}
\usage{
ensures_that(...)
}
\arguments{
\item{...}{conditions which must pass for the ensuring contract to be
fulfilled. Any named argument will treated as values available
when evaluating the conditions. To reference the value itself
use the dot-placeholder, \code{`.`}. The special name \code{fail_with}
is reserved for modifying behaviour on failure.}
}
\value{
an ensuring function
}
\description{
Make a contract which ensures that certain conditions are fulfilled.
}
\examples{
\dontrun{
library(magrittr)

# Ensures that a matrix is square.
ensure_square <- ensures_that(NCOL(.) == NROW(.))

diag(4) \%>\%
  ensure_square
}
}

