% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.ensurer}
\alias{print.ensurer}
\title{Print method for ensurer contracts}
\usage{
\method{print}{ensurer}(x)
}
\arguments{
\item{x}{a function made with \code{ensures_that}}
}
\value{
x
}
\description{
Print method for ensurer contracts
}

